Adding the Parameter Terrain to FieldHolder - GetTerrain - SetTerrain
adding a terrain grid to the gridWrapper to keep track of the grid.
added getSetTerrain to set the 2d terrain grid

added getGridCell into Gameboard to allow for easy gridcell access

trying to initialize all cells to Rocky Terrain. GameBoardBuilder line 105. and line 85

testing the getTerrain within the moveCommand.java line 70

added large value from 1 to 10 to test movement in RockyTerrain

changed commands to not take in game but now take in tank.